refactor: move formatDateTime functions to separated file
